+++ src/pointofsale/dao/implement/BillTableProductTmpDaoImpl.java
@@ -50,7 +50,7 @@ public class BillTableProductTmpDaoImpl extends SqlConstructor implements pointo
             statement.setInt(1, a.getBill_tmp_id());
             statement.setInt(2, a.getProduct_id());
             statement.setInt(3, a.getQuantity());
-            statement.setDouble(4, a.getSubvalue());
+            statement.setInt(4, a.getSubvalue());
             statement.executeUpdate();
             ResultSet idKey = statement.getGeneratedKeys();
             if (idKey.next()) {
@@ -98,7 +98,7 @@ public class BillTableProductTmpDaoImpl extends SqlConstructor implements pointo
             statement.setInt(1, a.getBill_tmp_id());
             statement.setInt(2, a.getProduct_id());
             statement.setInt(3, a.getQuantity());
-            statement.setDouble(4, a.getSubvalue());
+            statement.setInt(4, a.getSubvalue());
             statement.setInt(5, a.getId());
             if (statement.executeUpdate() == 0) {
                 System.out.println("Execute error");
@@ -175,7 +175,7 @@ public class BillTableProductTmpDaoImpl extends SqlConstructor implements pointo
         Integer bill_tmp_id = set.getInt("bill_tmp_id");
         Integer product_id = set.getInt("product_id");
         Integer quantity = set.getInt("quantity");
-        Double subvalue = set.getDouble("subvalue");
+        Integer subvalue = set.getInt("subvalue");
         String created_at = set.getString("created_at");
         BillTableProductTmp billTableProduct = new BillTableProductTmp(set.getInt("id"), bill_tmp_id, product_id, quantity, subvalue, created_at);
         return billTableProduct;
